:package: add base configuration packages

+++ webpack.mix.js
@@ -1,24 +1,25 @@
-const mix = require('laravel-mix')
-const path = require('path')
-const purgeCss = require('laravel-mix-purgecss')
+const mix = require('laravel-mix');
 
-mix
-  .js('resources/js/app.js', 'public/js')
-  .postCss('resources/css/app.css', 'public/css', [
-    require('postcss-import'),
-    require('tailwindcss')('./tailwind.config.js'),
-    require('postcss-nesting'),
-    require('autoprefixer'),
-  ])
-  .webpackConfig({
-    output: {chunkFilename: 'js/[name].js?id=[chunkhash]'},
-    resolve: {
-      alias: {
-        vue$: 'vue/dist/vue.runtime.esm.js',
-        '@': path.resolve('resources/js'),
-      },
-    },
-  })
-  .purgeCss()
-  .version()
-  .sourceMaps()
+require('laravel-mix-tailwind');
+require('laravel-mix-purgecss');
+
+/*
+ |--------------------------------------------------------------------------
+ | Mix Asset Management
+ |--------------------------------------------------------------------------
+ |
+ | Mix provides a clean, fluent API for defining some Webpack build steps
+ | for your Laravel application. By default, we are compiling the Sass
+ | file for the application as well as bundling up all the JS files.
+ |
+ */
+
+mix.js('resources/js/app.js', 'public/js')
+   .postCss('resources/css/app.css', 'public/css')
+   .tailwind('./tailwind.config.js');
+
+if (mix.inProduction()) {
+  mix
+   .version()
+   .purgeCss();
+}
